Add option to mute System.out and System.err.
The output of a test to `System.out` and `System.err` does not have to
be visible usually. Sometimes it may even slowdown the test.

The two new rules `SystemOutRule` and `SystemErrRule` support muting
these two streams.

    @Rule
    public final SystemOutRule systemOutRule = new SystemOutRule().mute();

    @Rule
    public final SystemErrRule systemErrRule = new SystemErrRule().mute();

`mute()` can be called within a test in order to mute the streams for
individual tests or parts of a test only.

The two new rules support logging, too. Logging is enabled by calling
`enableLogging()`. The log can be accessed by `getLog()`. It is
possible to clear the log during a test by calling `clearLog()`. These
functionality was already available from `StandardOutputStreamLog` and
`StandardErrorStreamLog`, which are now deprecated.